Implemented arpeggio effect (with refactoring)
-Arpeggio is now executed
-Audio generator now sends tick position to channel audio generator, which uses it to calculate arpeggio position
-When a finetune is specified, it is calculated in the channel audio genrator rather than the resampler

+++ src/main/kotlin/model/Row.kt
@@ -2,7 +2,7 @@ package model
 
 data class Row(
     val instrumentNumber: Int,
-    val period: Int,
+    val period: Float,
     val effect: EffectType,
     val effectXValue: Int,
     val effectYValue: Int
